Added first performance/correctness test. Start with: bin/gray -p perft_1

+++ inc/board_base.h
@@ -275,7 +275,7 @@ public:
 	virtual bool unmake();
 	virtual move_t san_to_coord(string& san);
 	virtual void coord_to_san(move_t m, string& san);
-	virtual int perft(int depth);
+	virtual unsigned long perft(int depth);
 
 protected:
 	list<state_t> states;                           // Previous states.
